start removing the android studio memory tooling
